UI updates and discount rate flexibility
Added flexibility to use discount rate from midpoint of cycle. Changed age and % female to be autonumericInputs (rather than numeric) to avoid the model crashing when invalid values are entered. Made the revised UI look a bit nicer. Updated the reference list to specify that the September 2021 version of the 2017-2019 England life table is being used.